backend/requirements.txt
orjson==3.10.0
typer[all]==0.6.1
tortoise-orm[asyncpg]==0.19.2
httpx==0.27.0
aiohttp==3.8.5
pydantic==1.10.7
sentry-sdk[asyncio,httpx]>=1.40.0
redis==5.0.0
python-telegram-bot==22.6
psycopg2-binary==2.9.9
nats-py==2.12.0
fastapi==0.104.1
uvicorn[standard]==0.30.1
segno==1.6.1
PyJWT>=2.8.0
